# public/docs/openapi/openAPI.yml
openapi: 3.0.0
servers:
  - url: https://paywithpretendpointsapi.onrender.com
info:
  version: 1.0.0
  title: Pay With Pretend Points API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/loyalty/members/{memberId}:
    get:
      tags:
        - Member Operations
      summary: Get a specific member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  member:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      name:
                        type: string
                        pattern: ^[\w\s]+$
                        example: John Doe
                      address:
                        type: string
                        pattern: ^[\w\s.,-]+$
                        example: 123 Main St, Toronto, ON
                      phone:
                        type: string
                        pattern: ^[\w\s.()-]+$
                        example: '4161234567'
                      email:
                        type: string
                        format: email
                        example: member@example.com
                      balance:
                        type: integer
                        example: 1000
                      status:
                        type: string
                        enum: *a3
                        default: PENDING
                        example: PENDING
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - partnerId
                      - name
                      - address
                      - phone
                      - email
                      - balance
                      - createdAt
                      - updatedAt
                required:
                  - member
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    patch:
      tags:
        - Member Operations
      summary: Update a specific member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  pattern: ^[\w\s]+$
                  example: John Doe
                address:
                  type: string
                  pattern: ^[\w\s.,-]+$
                  example: 123 Main St, Toronto, ON
                phone:
                  type: string
                  pattern: ^[\w\s.()-]+$
                  example: '4161234567'
                email:
                  type: string
                  format: email
                  example: member@example.com
                balance:
                  type: integer
                  example: 1000
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  member:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      name:
                        type: string
                        pattern: ^[\w\s]+$
                        example: John Doe
                      address:
                        type: string
                        pattern: ^[\w\s.,-]+$
                        example: 123 Main St, Toronto, ON
                      phone:
                        type: string
                        pattern: ^[\w\s.()-]+$
                        example: '4161234567'
                      email:
                        type: string
                        format: email
                        example: member@example.com
                      balance:
                        type: integer
                        example: 1000
                      status:
                        type: string
                        enum: *a3
                        default: PENDING
                        example: PENDING
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - partnerId
                      - name
                      - address
                      - phone
                      - email
                      - balance
                      - createdAt
                      - updatedAt
                required:
                  - member
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    delete:
      tags:
        - Member Operations
      summary: Delete a specific member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 1
                required:
                  - count
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
  /api/v1/loyalty/members:
    get:
      tags:
        - Member Operations
      summary: Get all members.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          minimum: 0
                          exclusiveMinimum: true
                          example: 1
                        partnerId:
                          type: integer
                          minimum: 0
                          exclusiveMinimum: true
                          example: 1
                        name:
                          type: string
                          pattern: ^[\w\s]+$
                          example: John Doe
                        address:
                          type: string
                          pattern: ^[\w\s.,-]+$
                          example: 123 Main St, Toronto, ON
                        phone:
                          type: string
                          pattern: ^[\w\s.()-]+$
                          example: '4161234567'
                        email:
                          type: string
                          format: email
                          example: member@example.com
                        balance:
                          type: integer
                          example: 1000
                        status:
                          type: string
                          enum: *a3
                          default: PENDING
                          example: PENDING
                        createdAt:
                          type: string
                          example: 2024-09-01T01:03:43.004Z
                        updatedAt:
                          type: string
                          example: 2024-09-01T01:03:43.004Z
                      required:
                        - id
                        - partnerId
                        - name
                        - address
                        - phone
                        - email
                        - balance
                        - createdAt
                        - updatedAt
                required:
                  - members
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    post:
      tags:
        - Member Operations
      summary: Create a new member.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  pattern: ^[\w\s]+$
                  example: John Doe
                address:
                  type: string
                  pattern: ^[\w\s.,-]+$
                  example: 123 Main St, Toronto, ON
                phone:
                  type: string
                  pattern: ^[\w\s.()-]+$
                  example: '4161234567'
                email:
                  type: string
                  format: email
                  example: member@example.com
                balance:
                  type: integer
                  example: 1000
              required:
                - name
                - address
                - phone
                - email
                - balance
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  member:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      name:
                        type: string
                        pattern: ^[\w\s]+$
                        example: John Doe
                      address:
                        type: string
                        pattern: ^[\w\s.,-]+$
                        example: 123 Main St, Toronto, ON
                      phone:
                        type: string
                        pattern: ^[\w\s.()-]+$
                        example: '4161234567'
                      email:
                        type: string
                        format: email
                        example: member@example.com
                      balance:
                        type: integer
                        example: 1000
                      status:
                        type: string
                        enum: *a3
                        default: PENDING
                        example: PENDING
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - partnerId
                      - name
                      - address
                      - phone
                      - email
                      - balance
                      - createdAt
                      - updatedAt
                required:
                  - member
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
...